+++ tic-tac-toe/src/app/components/board/index.js
@@ -1,50 +1,14 @@
 import React, { Component } from 'react';
+import PropTypes from 'prop-types';
 
 import Square from '../square';
 
 import style from './styles.scss';
 
 class Board extends Component {
-  state = {
-    squares: Array(9).fill(null),
-    xIsNext: true
-  };
-
-  handleClick = i => {
-    const squares = this.state.squares.slice();
-    if (this.calculateWinner(squares) || squares[i]) {
-      return;
-    }
-    squares[i] = this.state.xIsNext ? 'X' : 'O';
-    this.setState({
-      squares,
-      xIsNext: !this.state.xIsNext
-    });
-  };
-
-  calculateWinner = squares => {
-    const lines = [[0, 1, 2], [3, 4, 5], [6, 7, 8], [0, 3, 6], [1, 4, 7], [2, 5, 8], [0, 4, 8], [2, 4, 6]];
-
-    for (let i = 0; i < lines.length; i += 1) {
-      const [a, b, c] = lines[i];
-      if (squares[a] && squares[a] === squares[b] && squares[a] === squares[c]) {
-        return squares[a];
-      }
-    }
-    return null;
-  };
-
-  renderSquare = i => <Square value={this.state.squares[i]} onClick={() => this.handleClick(i)} />;
+  renderSquare = i => <Square value={this.props.squares[i]} onClick={() => this.props.onClick(i)} />;
 
   render() {
-    const winner = this.calculateWinner(this.state.squares);
-    let status;
-    if (winner) {
-      status = `Winner: ${winner}`;
-    } else {
-      status = `Next player:  ${this.state.xIsNext ? 'X' : 'O'}`;
-    }
-
     return (
       <React.Fragment>
         <div className={style.status}>{status}</div>
@@ -67,5 +31,10 @@ class Board extends Component {
     );
   }
 }
+
+Board.propTypes = {
+  squares: PropTypes.arrayOf(PropTypes.string),
+  onClick: PropTypes.func
+};
 
 export default Board;
